wip shutdown

+++ InMoov/life/shutdown.py
@@ -8,22 +8,8 @@
 
 
 def shutdown():
-  if isChatbotActivated:
-    chatBot.savePredicates()
-    chatBot.writeAndQuit()
-    RemoveFile(mrlWasNotCleanlyShutdownedFile)
+  if isChatbotActivated:RemoveFile(mrlWasNotCleanlyShutdownedFile)
   switchOnAllNervo()
-  talk(lang_shutDown)
   StopNeopixelAnimation()
   if isOpenCvActivated:opencv.stopCapture()
   if isKinectActivated:openni.stopCapture()
-  i01.halfSpeed()
-  i01.rest()
-  if isEyeLidsActivated:
-    eyelids.autoBlink(False)
-    eyelids.eyelidleft.moveTo(180)
-    eyelids.eyelidright.moveTo(180)
-  sleep(7)
-  i01.disable()
-  switchOffAllNervo()
-  killRuntime()
